requirements.txt
nest_asyncio==1.5.6
python-telegram-bot==20.1
langdetect==1.0.9
deep-translator==1.11.4
google-generativeai==0.7.2
flask==2.2.3
